Added new test in CopyFileTest.php for testing rollback

+++ tests/Actions/CopyFileTest.php
@@ -7,7 +7,6 @@ use \FSTransactions\Action\CopyFile;
 class CopyFileTest extends TestCase
 {
 
-
   private $destinationFilePath;
 
   public function setUp() {
@@ -17,13 +16,20 @@ class CopyFileTest extends TestCase
 
   public function tearDown() {
 
-    unlink($this->destinationFilePath);
+    if (file_exists($this->destinationFilePath)) {
+      unlink($this->destinationFilePath);
+    }
   }
 
+  /** 
+   * Test successfuly copying a file
+   * The file should exist in the destination
+   * and the contents should be the same
+   */
   public function testCopyFileExecuteSuccess() {
 
     $sourceFilePath = __DIR__.'/../test_files/CopyFileTestSource.txt';
-    $this->destinationFilePath = __DIR__.'/../test_files/CopyTestDestination.txt';
+    $this->destinationFilePath = __DIR__.'/../test_files/CopyTestTestDestination.txt';
 
     $action = new Copyfile($sourceFilePath, $this->destinationFilePath);
     $action->execute();
@@ -36,7 +42,22 @@ class CopyFileTest extends TestCase
     $this->assertEquals($sourceFileContents, $destinationFileContents);
   }
 
+  /**
+   * Test rolling back copying the file
+   * The copied file should be removed
+   */
   public function testCopyFileReverseSuccess() {
     
+    $sourceFilePath = __DIR__.'/../test_files/CopyFileTestSource.txt';
+    $this->destinationFilePath = __DIR__.'/../test_files/CopyTestTestDestination.txt';
+
+    $action = new Copyfile($sourceFilePath, $this->destinationFilePath);
+    $action->execute();
+
+    $this->assertFileExists($this->destinationFilePath);
+
+    $action->reverse();
+
+    $this->assertFileNotExists($this->destinationFilePath);
   }
 }
